+++ src/language.rs
@@ -64,7 +64,7 @@ impl Language {
     pub(crate) fn get_extension(language_name: &str, languages: &HashMap<String, Self>) -> String {
         let default = Self::default();
         if let Some(l) = languages.get(language_name) {
-            l.extension.to_owned()
+            l.extension.clone()
         } else {
             eprintln!(
                 "Couldn't find language {} in the list of extensions, defaulting to .txt",
@@ -92,8 +92,7 @@ pub fn get_languages(yml_string: &str) -> color_eyre::Result<HashMap<String, Lan
     let mut name_to_language = HashMap::new();
     for (name, language_yml) in languages {
         if let Some(extension) = language_yml.extensions.first() {
-            let mut language =
-                Language::new(name.to_owned(), extension.to_owned(), language_yml.color)?;
+            let mut language = Language::new(name.clone(), extension.clone(), language_yml.color)?;
             name_to_language.insert(name.to_ascii_lowercase(), language.clone());
             name_to_language.insert(name, language.clone());
             for alias in language_yml.aliases {
@@ -152,7 +151,7 @@ impl CodeHighlight {
         let mut theme_set = ThemeSet::load_defaults();
         theme_set
             .add_from_folder(&syntect_dir)
-            .map_err(|_| LostTheWay::ThemeError {
+            .map_err(|_e| LostTheWay::ThemeError {
                 theme: String::from((&syntect_dir).to_str().unwrap()),
             })
             .suggestion(format!(
@@ -162,7 +161,7 @@ impl CodeHighlight {
         let mut syntax_set = SyntaxSet::load_defaults_newlines().into_builder();
         syntax_set
             .add_from_folder(&syntect_dir, true)
-            .map_err(|_| LostTheWay::ThemeError {
+            .map_err(|_e| LostTheWay::ThemeError {
                 theme: String::from((&syntect_dir).to_str().unwrap()),
             })
             .suggestion(format!(
@@ -272,7 +271,7 @@ impl CodeHighlight {
 
     /// Gets current theme name
     pub(crate) fn get_theme_name(&self) -> String {
-        self.theme_name.to_owned()
+        self.theme_name.clone()
     }
 
     /// Adds a new theme from a .tmTheme file.
@@ -280,7 +279,7 @@ impl CodeHighlight {
     // TODO: should it automatically be set?
     pub(crate) fn add_theme(&mut self, theme_file: &Path) -> color_eyre::Result<String> {
         let theme = ThemeSet::get_theme(&theme_file)
-            .map_err(|_| LostTheWay::ThemeError {
+            .map_err(|_e| LostTheWay::ThemeError {
                 theme: theme_file.to_str().unwrap().into(),
             })
             .suggestion(format!(
@@ -289,7 +288,7 @@ impl CodeHighlight {
             ))?;
         let basename = theme_file
             .file_stem()
-            .and_then(|x| x.to_str())
+            .and_then(std::ffi::OsStr::to_str)
             .ok_or(LostTheWay::ThemeError {
                 theme: theme_file.to_str().unwrap().into(),
             })
@@ -309,7 +308,7 @@ impl CodeHighlight {
             true,
             None,
         )
-            .map_err(|_| LostTheWay::SyntaxError {
+            .map_err(|_e| LostTheWay::SyntaxError {
                 syntax: syntax_file.to_str().unwrap().into(),
             })
             .suggestion(format!(
@@ -318,7 +317,7 @@ impl CodeHighlight {
             ))?;
         let filename = syntax_file
             .file_name()
-            .and_then(|x| x.to_str())
+            .and_then(std::ffi::OsStr::to_str)
             .ok_or(LostTheWay::SyntaxError {
                 syntax: syntax_file.to_str().unwrap().into(),
             })
@@ -355,7 +354,7 @@ impl CodeHighlight {
             colorized.extend(
                 h.highlight(line, &self.syntax_set)
                     .into_iter()
-                    .map(|(style, s)| (style, s.to_string())),
+                    .map(|(style, s)| (style, s.to_owned())),
             );
         }
         colorized